Refactor: start in background instead of single Run
This is a big change of how the Service API works.

Old approach:

- itsy.New creates a Service instance
- The handlers can be registered before it is started
- Run start the service and only returns once the contact closes

New approach:

- itsy.Start creates a Service and starts it in the background
- The handlers can be registered any moment after the service was
  started
- You can call Stop to stop the service.

The new approach align better with how the NATS Micro API works. It
allows more flexibility, like dynamically adding handlers if needed, and
it makes it possible to provide other receivers that can be called at
runtime. For example, you can now easily access the NATS connection for
other API calls. This also removes the need for an InitFunc.

It is not ideal that there is window between the service registration
and the handler registration. Unfortunately the NATS Micro API does not
provide a better way. In practise the time between start and handler
registration should be short enough not to cause issues.